% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plane.R
\name{regression_plane}
\alias{regression_plane}
\title{Visualize a multiple regression model's predictions as a 3D plane}
\usage{
regression_plane(model, n_points = 100, mesh = FALSE, mesh_step = 1)
}
\arguments{
\item{model}{A fitted model object, e.g., the output from the lm() function.}

\item{n_points}{Integer, the number of steps from minimum to maximum value for which predicted values are generated from the model}

\item{mesh}{Logical, controls whether the regression plane has a mesh grid overlain on it}

\item{mesh_step}{step size for generating mesh grid lines}
}
\value{
a plotly plot object
}
\description{
Visualize a multiple regression model's predictions as a 3D plane
}
\examples{
m <- lm(mpg ~ wt + qsec, data = mtcars)
regression_plane(m)

}
